requirements.txt
fastapi
pandas
streamlit
ultralytics
Pillow
numpy
opencv-python
torch
plotly
